Fix expression for H

+++ Ising/src/accessors.cpp
@@ -24,7 +24,7 @@ uSize Ising::getSize() {
 	return this->p.N;
 }
 
-double Ising::getTemp() {
+float Ising::getTemp() {
 	return this->T;
 }
